requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
asyncio==3.4.3
dask[complete]==2023.12.1
numpy==1.24.4
pandas==2.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
xarray==2023.12.0